Poprawa warningu kompilacji

+++ src/src/sensor/VindriktningIkea.h
@@ -28,8 +28,8 @@
 #define VINDRIKTNING_IKEA_BAUDRATE 9600
 
 struct particleSensorState_t {
-  uint16_t avgPM25 = TEMPERATURE_NOT_AVAILABLE;
-  uint16_t measurements[5] = {0, 0, 0, 0, 0};
+  double avgPM25 = TEMPERATURE_NOT_AVAILABLE;
+  double measurements[5] = {0, 0, 0, 0, 0};
   uint8_t measurementIdx = 0;
   boolean valid = false;
 };
